+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## master
 
 ### Breaking Changes
+- vcl-input:
+  - Input 'value' is removed, use ngModel like you would with a regular input
+  - removed event-emitters since they are not needed
 - vcl-flip-switch:
   - switched to 2-way-binding
 - vcl-button:
@@ -24,13 +27,13 @@
 ### Breaking Changes
 - vcl-month-picker:
   - removed `@Input() useShortNames: boolean`, month names now default to short names, to use long month names, see `dateOptions` changes (have to pass in `{ month: 'long' }`)
-  
+
 ### Changes
 - vcl-month-picker:
   - introduced `@Input() locales: input: string | string[]`, see [Date.toLocaleDateString()](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/toLocaleDateString) first parameter
   - introduced `@Input() dateOptions: any`, see [Date.toLocaleDateString()](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/toLocaleDateString) second parameter
- 
-## 2017-01-17 - e3c8f1f 
+
+## 2017-01-17 - e3c8f1f
 
 ### Breaking Changes
 - vcl-dropdown: see [readme](https://github.com/ng-vcl/ng-vcl/blob/master/src/components/dropdown/README.md)